+++ src/main.hpp
@@ -58,7 +58,7 @@ struct peripherals_t {
 
 /* Prototypes */
 void init(struct peripherals_t &peripherals);
-uint16_t calculate_error(ccd_buffer_t &ccd_data);
+int calculate_error(ccd_buffer_t &ccd_data);
 void print_scan_result(struct peripherals_t &peripherals, ccd_buffer_t& ccd_data);
 void print_error_pos(struct peripherals_t &peripherals, double error);
